python script for finding most popular champions
Make sure to have the game data .txt file -gotten from getGameData.py- in the same directory. This program will give you all the combinations of champions occurring bellow the minimum support. Minimum support is an int passed to the script. The last entry in the output text file is the most frequent combinations of champions.